split out ed25519_multisig conversions into separate files
