+++ api/src/convert/mint_config.rs
@@ -90,7 +90,9 @@ impl TryFrom<&external::MintConfigTx> for MintConfigTx {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::convert::ed25519_multisig::tests::{test_multi_sig, test_signer_set};
+    use crate::convert::ed25519_multisig::{
+        multisig::tests::test_multi_sig, signerset_v1::tests::test_signer_set_v1,
+    };
     use mc_util_serial::{decode, encode};
     use protobuf::Message;
 
@@ -100,7 +102,7 @@ mod tests {
     fn test_convert_mint_config() {
         let source = MintConfig {
             token_id: 123,
-            signer_set: test_signer_set(),
+            signer_set: test_signer_set_v1(),
             mint_limit: 10000,
         };
 
@@ -146,12 +148,12 @@ mod tests {
                 configs: vec![
                     MintConfig {
                         token_id: 123,
-                        signer_set: test_signer_set(),
+                        signer_set: test_signer_set_v1(),
                         mint_limit: 10000,
                     },
                     MintConfig {
                         token_id: 456,
-                        signer_set: test_signer_set(),
+                        signer_set: test_signer_set_v1(),
                         mint_limit: 20000,
                     },
                 ],
